Support block statements - kinda
It doesn't work very well because it doesn't capture local scope
within templates. It also leads to some really dirty looking templates.
It's possible it could work well if the code was processed by babel
or something and had all local variables captured in a way that
could safely be forwarded on to the context - but that seems like
quite a lot of work.